+++ aiida_restapi/routers/server.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import re
 
-import pydantic as pdt
 from aiida import __version__ as aiida_version
 from fastapi import APIRouter, Request
 from fastapi.responses import HTMLResponse
@@ -12,33 +11,9 @@ from fastapi.routing import APIRoute
 from starlette.routing import Route
 
 from aiida_restapi.config import API_CONFIG
+from aiida_restapi.models.server import ServerEndpoint, ServerInfo
 
 read_router = APIRouter(prefix='/server')
-
-
-class ServerInfo(pdt.BaseModel):
-    """API version information."""
-
-    api_major_version: str = pdt.Field(
-        description='Major version of the API',
-        examples=['0'],
-    )
-    api_minor_version: str = pdt.Field(
-        description='Minor version of the API',
-        examples=['1'],
-    )
-    api_revision_version: str = pdt.Field(
-        description='Revision version of the API',
-        examples=['0a1'],
-    )
-    api_prefix: str = pdt.Field(
-        description='Prefix for all API endpoints',
-        examples=['/api/v0'],
-    )
-    aiida_version: str = pdt.Field(
-        description='Version of the AiiDA installation',
-        examples=['2.8.0'],
-    )
 
 
 @read_router.get(
@@ -55,28 +30,6 @@ async def get_server_info() -> dict[str, str]:
         'api_prefix': API_CONFIG['PREFIX'],
         'aiida_version': aiida_version,
     }
-
-
-class ServerEndpoint(pdt.BaseModel):
-    """API endpoint."""
-
-    path: str = pdt.Field(
-        description='Path of the endpoint',
-        examples=['../server/endpoints'],
-    )
-    group: str | None = pdt.Field(
-        description='Group of the endpoint',
-        examples=['server'],
-    )
-    methods: set[str] = pdt.Field(
-        description='HTTP methods supported by the endpoint',
-        examples=['GET'],
-    )
-    description: str = pdt.Field(
-        '-',
-        description='Description of the endpoint',
-        examples=['Get a JSON-serializable dictionary of all registered API routes.'],
-    )
 
 
 @read_router.get(
